audio match int16_t

+++ include/VoiceDetector.h
@@ -1,5 +1,5 @@
 // ============================================================================
-// VoiceDetector.h - Updated for (79, 10, 1) input
+// VoiceDetector.h - int16_t only
 // ============================================================================
 #ifndef VOICE_DETECTOR_H
 #define VOICE_DETECTOR_H
@@ -17,8 +17,8 @@ public:
     VoiceDetector();
     ~VoiceDetector();
     
-    // Process audio and return detection score (0.0 to 1.0)
-    float detectWakeWord(const float* audio, int length);
+    // Process int16_t audio (matches ESP32 mic format)
+    float detectWakeWord(const int16_t* audio, int length);
     
     // Helper to print MFCC features for debugging
     void printMFCC(int frame);
